Урок 5: Асинхронный вызов через RabbitMQ. Вычисление эмбеддингов теперь выполняется асинхронно. Эмбеддинги через реальную ML модель

See merge request konstantin-lebedev-fwf6667/semanticsearch!5

+++ common/models/query_result_item.py
@@ -1,13 +1,14 @@
 from __future__ import annotations
 
 from uuid import uuid4
+from decimal import Decimal
 
 from sqlalchemy import ForeignKey, Integer, Numeric, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from database.database import Base
-from models.mixins import CrudMixin
+from common.database.database import Base
+from common.models.mixins import CrudMixin
 
 
 class QueryResultItem(Base, CrudMixin):
@@ -32,7 +33,7 @@ class QueryResultItem(Base, CrudMixin):
         nullable=False,
     )
 
-    score: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False)
+    score: Mapped[Decimal] = mapped_column(Numeric(10, 6), nullable=False)
     rank: Mapped[int] = mapped_column(Integer, nullable=False)
 
     query: Mapped["Query"] = relationship(back_populates="results")
